+++ src/mailman/commands/tests/test_cli_digests.py
@@ -347,6 +347,115 @@ Subject: message 1
         items = get_queue_messages('virgin', expected_count=1)
         self.assertEqual(items[0].msg['subject'], 'Ant Digest, Vol 8, Issue 1')
 
+    def test_send_periodic_one_by_listid(self):
+        # Test sending digest using --periodic.
+        msg = mfs("""\
+To: ant@example.com
+From: anne@example.com
+Subject: message 1
+
+""")
+        self._handler.process(self._mlist, msg, {})
+        del msg['subject']
+        msg['subject'] = 'message 2'
+        self._handler.process(self._mlist, msg, {})
+        # There are no digests already being sent, but the ant mailing list
+        # does have a digest mbox collecting messages.
+        get_queue_messages('digest', expected_count=0)
+        mailbox_path = os.path.join(self._mlist.data_path, 'digest.mmdf')
+        self.assertGreater(os.path.getsize(mailbox_path), 0)
+        self._command.invoke(digests, ('-p', '-v', '-l', 'ant.example.com'))
+        self._runner.run()
+        # Now, there's no digest mbox and there's a plaintext digest in the
+        # outgoing queue.
+        self.assertFalse(os.path.exists(mailbox_path))
+        items = get_queue_messages('virgin', expected_count=1)
+        digest_contents = str(items[0].msg)
+        self.assertIn('Subject: message 1', digest_contents)
+        self.assertIn('Subject: message 2', digest_contents)
+
+    def test_send_periodic_set_false(self):
+        # Test sending digest --periodic when the only Mailing List's
+        # digest_send_periodic is set to false.
+        # Test sending digest using --periodic.
+        self._mlist.digest_send_periodic = False
+        msg = mfs("""\
+To: ant@example.com
+From: anne@example.com
+Subject: message 1
+
+""")
+        self._handler.process(self._mlist, msg, {})
+        # There are no digests already being sent, but the ant mailing list
+        # does have a digest mbox collecting messages.
+        get_queue_messages('digest', expected_count=0)
+        mailbox_path = os.path.join(self._mlist.data_path, 'digest.mmdf')
+        self.assertGreater(os.path.getsize(mailbox_path), 0)
+        self._command.invoke(digests, ('-p', '-l', 'ant.example.com'))
+        self._runner.run()
+        # Now, even though the digest command was run, the mailbox should still
+        # be there.
+        self.assertTrue(os.path.exists(mailbox_path))
+        get_queue_messages('virgin', expected_count=0)
+
+    def test_send_periodic_two_lists_one_set_false(self):
+        # Test digests --periodic when one of the two lists has
+        # digest_send_periodic set to false.
+        # Populate ant's digest.
+        self._mlist.digest_send_periodic = False
+        msg = mfs("""\
+To: ant@example.com
+From: anne@example.com
+Subject: message 1
+
+""")
+        self._handler.process(self._mlist, msg, {})
+        del msg['subject']
+        msg['subject'] = 'message 2'
+        self._handler.process(self._mlist, msg, {})
+        # Create the second list.
+        bee = create_list('bee@example.com')
+        bee.digests_enabled = True
+        bee.digest_size_threshold = 100000
+        bee.send_welcome_message = False
+        member = subscribe(bee, 'Bart')
+        member.preferences.delivery_mode = DeliveryMode.plaintext_digests
+        # Populate bee's digest.
+        msg = mfs("""\
+To: bee@example.com
+From: bart@example.com
+Subject: message 3
+
+""")
+        self._handler.process(bee, msg, {})
+        del msg['subject']
+        msg['subject'] = 'message 4'
+        self._handler.process(bee, msg, {})
+        # There are no digests for either list already being sent, but the
+        # mailing lists do have a digest mbox collecting messages.
+        ant_mailbox_path = os.path.join(self._mlist.data_path, 'digest.mmdf')
+        self.assertGreater(os.path.getsize(ant_mailbox_path), 0)
+        # Check bee's digest.
+        bee_mailbox_path = os.path.join(bee.data_path, 'digest.mmdf')
+        self.assertGreater(os.path.getsize(bee_mailbox_path), 0)
+        # Both.
+        get_queue_messages('digest', expected_count=0)
+        # Process both list's digests.
+        self._command.invoke(
+            digests, ('-p', '-l', 'ant.example.com', '-l', 'bee@example.com'))
+        self._runner.run()
+        # Now, ant should still have it's mailbox file, but bee shouldn't.
+        # Also, bee's message should be in the outgoing queue.
+        self.assertTrue(os.path.exists(ant_mailbox_path))
+        self.assertFalse(os.path.exists(bee_mailbox_path))
+        items = get_queue_messages('virgin', expected_count=1)
+        # Figure out which digest is going to ant and which to bee.
+        assert items[0].msg['to'] == 'bee@example.com'
+        # Check bee's digest.
+        digest_contents = str(items[0].msg)
+        self.assertIn('Subject: message 3', digest_contents)
+        self.assertIn('Subject: message 4', digest_contents)
+
 
 class TestBumpVolume(unittest.TestCase):
     layer = ConfigLayer
@@ -396,4 +505,20 @@ ant.example.com bumped to volume 7, number 5
             digests, ('-v', '-s', '-n', '-l', 'ant.example.com'))
         self.assertMultiLineEqual(result.output, """\
 ant.example.com sent volume 7, number 4
+""")
+
+
+class TestDigestCommand(unittest.TestCase):
+    layer = ConfigLayer
+
+    def setUp(self):
+        self._mlist = create_list('ant@example.com')
+        self._command = CliRunner()
+
+    def test_send_and_periodic_options_exclusive(self):
+        result = self._command.invoke(digests,
+                                      ('-s', '-p', '-l', 'ant@example.com'))
+        self.assertEqual(result.exit_code, 1)
+        self.assertMultiLineEqual(result.output, """\
+--send and --periodic flags cannot be used together
 """)